TMSnapshot Query
Scope job used for pulling dashboard data. Dashboard displays data for a specific purchase scenario and population for the last 10 days and features tiles that display performance for the last 24 hours. Data for the last 24 hours is indicated by the IsRealTime field with a value of TRUE. Purchase scenario and customer population is pulled via an INNER JOIN to another structured stream that contains specific BINs.