Job_Automation
PS script for automating the running and exporting of a cosmos/scope job. Script submits the scope job and sends an email notification to the specified email address. Upon successful completion of the job, the structured stream is exported to a cloud location stored CSV where it is used to populate a PowerBI dashboard.